+++ CHANGES.rst
@@ -4,6 +4,9 @@ Changelog
 0.1-dev (unreleased)
 --------------------
 
+- Ensure users logging in get the Shibboleth Authenticated role via
+  AuthZ mapping.
+  [davidjb]
 - Add Shibboleth Authenticated role.
   [davidjb]
 - Package created using templer